优化格式化对象

+++ audioStation/src/main/java/com/faker/audioStation/model/dto/wyy/songDetail/ChargeInfoList.java
@@ -18,7 +18,7 @@ import java.io.Serializable;
 @NoArgsConstructor
 public class ChargeInfoList implements Serializable {
 
-    private long rate;
-    private int chargeType;
+    private Long rate;
+    private Long chargeType;
 
 }
